保存采购单信息

+++ oms_commodity/src/main/java/com/oms/saas/commodity/Entity/Warehouse/WmsTickets.java
@@ -246,5 +246,5 @@ public class WmsTickets implements Serializable {
      */
     private Date acceptCallbackTime;
 
-    private Integer actualWarehouse;
+    private Byte actualWarehouse;
 }
